Extract Scope, SelfSymbolImpl and SymbolImpl classes from SymbolTableBuilder
